+++ exporters/trace/zipkin/model.go
@@ -26,6 +26,11 @@ import (
 	export "go.opentelemetry.io/otel/sdk/export/trace"
 )
 
+const (
+	keyInstrumentationLibraryName    = "otel.instrumentation_library.name"
+	keyInstrumentationLibraryVersion = "otel.instrumentation_library.version"
+)
+
 func toZipkinSpanModels(batch []*export.SpanData, serviceName string) []zkmodel.SpanModel {
 	models := make([]zkmodel.SpanModel, 0, len(batch))
 	for _, data := range batch {
@@ -132,9 +137,16 @@ func attributesToJSONMapString(attributes []label.KeyValue) string {
 	return (string)(jsonBytes)
 }
 
+// extraZipkinTags are those that may be added to every outgoing span
+var extraZipkinTags = []string{
+	"ot.status_code",
+	"ot.status_description",
+	keyInstrumentationLibraryName,
+	keyInstrumentationLibraryVersion,
+}
+
 func toZipkinTags(data *export.SpanData) map[string]string {
-	// +2 for status code and for status message
-	m := make(map[string]string, len(data.Attributes)+2)
+	m := make(map[string]string, len(data.Attributes)+len(extraZipkinTags))
 	for _, kv := range data.Attributes {
 		m[(string)(kv.Key)] = kv.Value.Emit()
 	}
@@ -143,5 +155,12 @@ func toZipkinTags(data *export.SpanData) map[string]string {
 	}
 	m["ot.status_code"] = data.StatusCode.String()
 	m["ot.status_description"] = data.StatusMessage
+
+	if il := data.InstrumentationLibrary; il.Name != "" {
+		m[keyInstrumentationLibraryName] = il.Name
+		if il.Version != "" {
+			m[keyInstrumentationLibraryVersion] = il.Version
+		}
+	}
 	return m
 }